Restructure /auth routes
