+++ apps/backend/src/routes/index.ts
@@ -1,0 +1,8 @@
+import * as express from 'express';
+import authRouter from './auth';
+
+const router = express.Router();
+
+router.use('/auth', authRouter);
+
+export default router;
